+++ apworld/brotato/__init__.py
@@ -181,16 +181,20 @@ class BrotatoWorld(World):
     def generate_early(self) -> None:
         # Determine needed values from the options
         self.waves_with_checks = get_waves_with_checks(self.options.waves_per_drop)
-
-        self._include_characters, self._starting_characters = get_available_and_starting_characters(
-            self.options.include_base_game_characters.value,
-            bool(self.options.enable_abyssal_terrors_dlc.value),
-            self.options.include_abyssal_terrors_characters.value,
-            self.options.starting_characters,
-            self.options.num_starting_characters.value,
-            self.options.num_characters.value,
-            self.random,
-        )
+        re_gen_passthrough = getattr(self.multiworld, "re_gen_passthrough", {})
+        if re_gen_passthrough:
+            self._starting_characters = re_gen_passthrough[self.game]["starting_characters"]
+            self._include_characters = re_gen_passthrough[self.game]["include_characters"]
+        else:
+            self._include_characters, self._starting_characters = get_available_and_starting_characters(
+                self.options.include_base_game_characters.value,
+                bool(self.options.enable_abyssal_terrors_dlc.value),
+                self.options.include_abyssal_terrors_characters.value,
+                self.options.starting_characters,
+                self.options.num_starting_characters.value,
+                self.options.num_characters.value,
+                self.random,
+            )
 
         # Clamp the number of wins needed to goal to the number of included characters, so the game isn't unwinnable.
         self.num_wins_needed = min(self.options.num_victories.value, len(self._include_characters))
@@ -332,4 +336,18 @@ class BrotatoWorld(World):
             "legendary_crate_drop_groups": [asdict(g) for g in self.legendary_loot_crate_groups],
             "wave_per_game_item": wave_per_game_item,
             "enable_abyssal_terrors_dlc": self.options.enable_abyssal_terrors_dlc.value,
+            # Info for Universal Tracker, the mod doesn't use these.
+            "starting_characters": self._starting_characters,
+            "include_characters": self._include_characters,
         }
+
+    @staticmethod
+    def interpret_slot_data(slot_data: dict[str, Any]) -> dict[str, Any] | None:
+        if "starting_characters" in slot_data:
+            # If this key is present, all UT-related keys should be present
+            return {
+                "starting_characters": slot_data["starting_characters"],
+                "include_characters": slot_data["include_characters"],
+            }
+        # Backwards compat with worlds generated before UT support was added
+        return None