+++ CHANGES.md
@@ -7,6 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Reduced the maximum values of some options that were too unfun at larger values:
+  - `num_victories` is now 25 instead of matching the total number of characters.
+  - `num_common_crate_drops` is now 25 instead of 50.
+  - `num_legendary_crate_drops` is now 10 instead of 50.
+
 ## [0.15.2] - 2026-08-14
 
 ### Fixed