+++ CHANGES.md
@@ -10,12 +10,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - New Feature: Progressive Wave Caps
   - The game now has a cap on the highest wave that can be played from the start.
-  - A new item, "Progressive Wave Cap", has been added which, when received, increases
-    the maximum playable wave.
+  - A set of new items, "Progressive Wave Cap (<character>)", has been added which,
+    when received, increases the maximum playable wave for each character.
   - The logic for Wave Completed and Run Won locations now requires that the wave cap be
     greater than the wave number or 20, respectively.
   - Controlled by a new option, `wave_caps`, which sets the number Progressive Wave Cap
-    Increase items and the increase given by each one.
+    Increase items for each character and the increase given by each one.
   - There are predefined values for `wave_caps`:
     - `one`: One wave cap, all 20 waves are reachable from the start
       (matches previous apworld behavior).\